+++ src/app/demo/view/inputdemo.component.ts
@@ -91,6 +91,7 @@ export class InputDemoComponent implements OnInit{
 
     constructor(private countryService: CountryService, private breadcrumbService: BreadcrumbService) {
         this.breadcrumbService.setItems([
+            {label: 'UI Kit'},
             {label: 'Input'}
         ]);
     }