+++ src/app/demo/view/mediademo.component.ts
@@ -2,10 +2,10 @@ import {Component, OnInit} from '@angular/core';
 import { ProductService } from '../service/productservice';
 import { PhotoService } from '../service/photoservice';
 import { Product } from '../domain/product';
+import { BreadcrumbService } from 'src/app/app.breadcrumb.service';
 
 @Component({
     templateUrl: './mediademo.component.html',
-    styleUrls: ['./mediademo.scss'],
 })
 export class MediaDemoComponent implements OnInit{
 
@@ -50,7 +50,12 @@ export class MediaDemoComponent implements OnInit{
         }
     ];
 
-    constructor(private productService: ProductService, private photoService: PhotoService) {}
+    constructor(private productService: ProductService, private photoService: PhotoService, public breadcrumbService: BreadcrumbService) {
+        this.breadcrumbService.setItems([
+            {label: 'UI Kit'},
+            {label: 'Media'}
+        ]);
+    }
 
     ngOnInit() {
         this.productService.getProductsSmall().then(products => {