+++ src/app/pages/app.crud.component.ts
@@ -7,7 +7,6 @@ import {BreadcrumbService} from '../app.breadcrumb.service';
 
 @Component({
     templateUrl: './app.crud.component.html',
-    styleUrls: ['../demo/view/tabledemo.scss'],
     styles: [`
         :host ::ng-deep .p-dialog .product-image {
             width: 150px;
@@ -31,6 +30,10 @@ export class AppCrudComponent implements OnInit {
 
     productDialog: boolean;
 
+    deleteProductDialog: boolean = false;
+
+    deleteProductsDialog: boolean = false;
+
     products: Product[];
 
     product: Product;
@@ -41,22 +44,32 @@ export class AppCrudComponent implements OnInit {
 
     cols: any[];
 
+    statuses: any[];
+
+    rowsPerPageOptions = [5, 10, 20];
+
     constructor(private productService: ProductService, private messageService: MessageService,
                 private confirmationService: ConfirmationService, private breadcrumbService: BreadcrumbService) {
         this.breadcrumbService.setItems([
+            {label: 'Pages'},
             {label: 'Crud'}
         ]);
     }
-
-    ngOnInit() {
+   ngOnInit() {
         this.productService.getProducts().then(data => this.products = data);
 
         this.cols = [
-            { field: 'name', header: 'Name' },
-            { field: 'price', header: 'Price' },
-            { field: 'category', header: 'Category' },
-            { field: 'rating', header: 'Reviews' },
-            { field: 'inventoryStatus', header: 'Status' }
+            {field: 'name', header: 'Name'},
+            {field: 'price', header: 'Price'},
+            {field: 'category', header: 'Category'},
+            {field: 'rating', header: 'Reviews'},
+            {field: 'inventoryStatus', header: 'Status'}
+        ];
+
+        this.statuses = [
+            {label: 'INSTOCK', value: 'instock'},
+            {label: 'LOWSTOCK', value: 'lowstock'},
+            {label: 'OUTOFSTOCK', value: 'outofstock'}
         ];
     }
 
@@ -67,16 +80,7 @@ export class AppCrudComponent implements OnInit {
     }
 
     deleteSelectedProducts() {
-        this.confirmationService.confirm({
-            message: 'Are you sure you want to delete the selected products?',
-            header: 'Confirm',
-            icon: 'pi pi-exclamation-triangle',
-            accept: () => {
-                this.products = this.products.filter(val => !this.selectedProducts.includes(val));
-                this.selectedProducts = null;
-                this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Products Deleted', life: 3000});
-            }
-        });
+        this.deleteProductsDialog = true;
     }
 
     editProduct(product: Product) {
@@ -85,16 +89,22 @@ export class AppCrudComponent implements OnInit {
     }
 
     deleteProduct(product: Product) {
-        this.confirmationService.confirm({
-            message: 'Are you sure you want to delete ' + product.name + '?',
-            header: 'Confirm',
-            icon: 'pi pi-exclamation-triangle',
-            accept: () => {
-                this.products = this.products.filter(val => val.id !== product.id);
-                this.product = {};
-                this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Product Deleted', life: 3000});
-            }
-        });
+        this.deleteProductDialog = true;
+        this.product = {...product};
+    }
+
+    confirmDeleteSelected(){
+        this.deleteProductsDialog = false;
+        this.products = this.products.filter(val => !this.selectedProducts.includes(val));
+        this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Products Deleted', life: 3000});
+        this.selectedProducts = null;
+    }
+
+    confirmDelete(){
+        this.deleteProductDialog = false;
+        this.products = this.products.filter(val => val.id !== this.product.id);
+        this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Product Deleted', life: 3000});
+        this.product = {};
     }
 
     hideDialog() {
@@ -107,12 +117,16 @@ export class AppCrudComponent implements OnInit {
 
         if (this.product.name.trim()) {
             if (this.product.id) {
+                // @ts-ignore
+                this.product.inventoryStatus = this.product.inventoryStatus.value ? this.product.inventoryStatus.value: this.product.inventoryStatus;
                 this.products[this.findIndexById(this.product.id)] = this.product;
                 this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Product Updated', life: 3000});
-            }
-            else {
+            } else {
                 this.product.id = this.createId();
+                this.product.code = this.createId();
                 this.product.image = 'product-placeholder.svg';
+                // @ts-ignore
+                this.product.inventoryStatus = this.product.inventoryStatus ? this.product.inventoryStatus.value : 'INSTOCK';
                 this.products.push(this.product);
                 this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Product Created', life: 3000});
             }
@@ -138,7 +152,7 @@ export class AppCrudComponent implements OnInit {
     createId(): string {
         let id = '';
         const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-        for ( let i = 0; i < 5; i++ ) {
+        for (let i = 0; i < 5; i++) {
             id += chars.charAt(Math.floor(Math.random() * chars.length));
         }
         return id;