+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+
+## 13.1.0
+**Migration Guide**
+- Update to PrimFlex 3.
+- Update to Angular 13 and PrimeNG 13.
+
 ## 12.2.0
 **Migration Guide**
 - Update theme files and layout files.